requirements.txt
streamlit
transformers
torch
torchvision
pillow
requests
python-dotenv
